requirements.txt
pandas
scikit-learn
kafka-python
streamlit
faker
joblib
numpy
altair
psutil
streamlit-lottie
